Added all the node_modules folder so that this can be an npm package.
Support for github-flavored-markdown, via an npm package of the same name. All posts are being rendered correctly with this method, and without clobbering Syntax Highlighter.

Changed the db user to 'simpleblog'; I kind of want this to be an npm package.

Restructured postPrerenderer and renamed it to postToHTML. It can handle an single post or an array of them.

Slight css tweak to posts' dates.