Function::curry works perfectly!
You don't have to use Function::curry. nCurry and Currier are great to
use too by themselves, but Function::curry the nicest of them because it
accepts n args at any point in the currying (even in the call to
f.curry itself).

Vector2 two now has a Zero function. Using it in gravity.